% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{km_to_miles}
\alias{km_to_miles}
\title{Convert km to miles}
\usage{
km_to_miles(km)
}
\arguments{
\item{km}{Distance in km}
}
\value{
Distance in miles
}
\description{
This function interprets a numerical value in terms of the distance
unit kilometers and converts it to miles.
}
\examples{

km_to_miles(1)

}
\references{
\url{https://en.wikipedia.org/wiki/Mile}\cr
}
\seealso{
\code{\link{miles_to_km}}\cr
\code{\link{miles_to_km2}}\cr
}
\author{
Fateema Bazzi\cr Politecnico di Milano\cr Maintainer: Fateema
Bazzi\cr E-Mail: \href{mailto:fateemahani.bazzi@mail.polimi.it}{fateemahani.bazzi@mail.polimi.it}
}
