% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{miles_to_km2}
\alias{miles_to_km2}
\title{Convert miles to km}
\usage{
miles_to_km2(miles)
}
\arguments{
\item{miles}{Distance in miles}
}
\value{
Distance in kilometers
}
\description{
This function interprets a numerical value in terms of the distance
unit miles and converts it to kilometers.
}
\examples{

miles_to_km2(1)

}
\references{
\url{https://en.wikipedia.org/wiki/Mile}\cr
}
\seealso{
\code{\link{km_to_miles}}\cr
\code{\link{miles_to_km}}\cr
}
\author{
Fateema Bazzi\cr Politecnico di Milano\cr Maintainer: Fateema
Bazzi\cr E-Mail: \href{mailto:fateemahani.bazzi@mail.polimi.it}{fateemahani.bazzi@mail.polimi.it}
}
