% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distConvert-package.R
\docType{package}
\name{distConvert-package}
\alias{distConvert-package}
\alias{distConvert}
\title{distConvert - Convert distances from kilometres to miles and vice versa}
\description{
The \code{distConvert} package implements several functions for conversion
of different distance measures, namely kilometres and miles.
}
\details{
\tabular{ll}{
Package: \tab distConvert\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2020-04-15\cr
License: \tab GPL (>=2)\cr
}
}
\references{
\url{http://some.url/here/maybe/}\cr
Piro et al. (2007) Some paper. Journal, volume, issue.\cr
}
\author{
Rosario M. Piro [aut, cre]\cr
Politecnico di Milano\cr
Maintainer: Rosario M. Piro\cr
E-Mail: \href{mailto:rosariomichael.piro@polimi.it}{rosariomichael.piro@polimi.it}
}
